// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const InvoiceType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type InvoiceType = (typeof InvoiceType)[keyof typeof InvoiceType]


export const PriceSource = {
  AUTO: 'AUTO',
  MANUAL: 'MANUAL'
} as const

export type PriceSource = (typeof PriceSource)[keyof typeof PriceSource]


export const UnitType = {
  PIECE: 'PIECE',
  BOX: 'BOX'
} as const

export type UnitType = (typeof UnitType)[keyof typeof UnitType]


export const AgentInvoiceFormat = {
  STANDARD: 'STANDARD',
  BOX_IN_QTY: 'BOX_IN_QTY',
  CUSTOM_4: 'CUSTOM_4'
} as const

export type AgentInvoiceFormat = (typeof AgentInvoiceFormat)[keyof typeof AgentInvoiceFormat]